updated path to assets in the gym testing, and united the ddp and fsdp tests into a single `tests_distributed.py`
